requirements.txt
python-telegram-bot==20.0
Pillow==8.0.0
apscheduler==3.8.1
